operations: Add concept of OPERATIONS_REGISTER.
With this kinda registery what we hope to achieve is that every time
we add new operations which might be possible to apply to a
List[List[Response]] we will add it to this registery and using this
regitery the UI can provide options to attach these operations to a
form. Once attached we will be able to query for operations to apply
and execute their scripts.